app, plug-ins: (Win32) fix calls to ShellExecute().
This function returns a HINSTANCE (which is technically a pointer if I
understand correctly) for backward compatibility reason. But this is
actually meant to be used as an int only according to the docs.
Yet casting to gint outputs "cast from pointer to integer of different
size" warnings.
We should use intptr_t instead for such cases, since it is a type made
explicitly to hold a pointer as int, hence should always be right on any
platform.

https://docs.microsoft.com/en-us/windows/win32/api/shellapi/nf-shellapi-shellexecutea

+++ plug-ins/common/web-browser.c
@@ -165,11 +165,11 @@ browser_open_url (GtkWindow    *window,
   HINSTANCE hinst = ShellExecute (GetDesktopWindow(),
                                   "open", url, NULL, NULL, SW_SHOW);
 
-  if ((gint) hinst <= 32)
+  if ((intptr_t) hinst <= 32)
     {
       const gchar *err;
 
-      switch ((gint) hinst)
+      switch ((intptr_t) hinst)
         {
           case 0 :
             err = _("The operating system is out of memory or resources.");
